Remove FZF preview.
We use it most of the time in neovim, and previewing tags (for instance)
is undefined behavior (it gives errors we don’t wanna know about). That
was an interesting feature but we can enable it with a zsh alias if we
really want it (I actually don’t really give a heck).